requirements.txt
pandas
tqdm
sentence-transformers
numpy
faiss-cpu
gradio